pyproject.toml
[tool.poetry]
name = "rag-chatbot-backend"
version = "0.1.0"
description = ""
authors = ["Gopal Katariya <gopal.katariya@differenzsystem.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = "^0.27.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.6.1"}
pydantic-settings = "^2.2.1"
langchain = "^0.1.9"
langchain-openai = "^0.0.5"
langchain-google-genai = "^0.0.9"
pgvector = "^0.2.5"
cryptography = "^42.0.2"
langchain-community = "^0.0.24"
langchain-core = "^0.1.30"
pypdf = "^4.0.2"
python-docx = "^1.1.0"
unstructured = "^0.11.8"
psycopg2 = "^2.9.10"
ftfy = "^6.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
